Updated source without useless hardcoded currencies to resolve #254.
Free Software Free Society

To support commits by ctubio,
you can buy-me-a-drink with a small git tip at:
  1GitTipgxvKB3zjCLXRcSgDhC9pivkpc7u

I promise to drink chocolate milk in the development of the next commit.

To request new features or in case this commit breaks something for you,
please create a new github issue with all possible details,
but never share your API Keys!

+++ src/server/gw.h
@@ -64,8 +64,8 @@ namespace K {
     private:
       static json onSnapProduct() {
         return {{
-          {"exchange", (double)gw->exchange},
-          {"pair", {{"base", (double)gw->base}, {"quote", (double)gw->quote}}},
+          {"exchange", (int)gw->exchange},
+          {"pair", {{"base", gw->base}, {"quote", gw->quote}}},
           {"minTick", gw->minTick},
           {"environment", argTitle},
           {"matryoshka", argMatryoshka},
